// Fro042023/Semana 5/proyectoAngular/primerAngular/src/app/models/curso.ts
export class Curso {
    id: string = "";
    nombre: string = "";
    descripcion: string= "";
    tiempo: string= "";
    usuario: string= "";
}


//{"code":200,"message":"Cursos Lista",
// "data":
// [{"id":"3925",
// "nombre":"Prueba Jes\u00fas 3",
// "descripcion":"Prueba Jes\u00fas",
// "tiempo":"5 Semanas","usuario":"auguzman"},